Completed Off-Site Testing for a Carousel, as it continuously kept breaking without knowing.
Fixed Carousel, found an error with it overflowing
but the terminal could not find what was causing it

As it turns out, the images were, after resizing it
persisted, as a result, temporary hidden overflow
is added, though I believe it has to do with
the hamburger nav.

Things to do:
The Contact Page.
Finish the Home Page.
Products Page.

Header/Footer Redesign are Complete now.

Back End must be completed coming up, as soon
as the front end is rounded out into a functioning
rough draft.

JavaScript animations have also been added.